Use '#pragma once' instead of '#ifndef' guards
'#pragma once', while not part of the C++11 standard, is supported
across virtually all compilers, including g++ and clang++.

In this case, the '#ifndef SCOPE' guard in scope.hpp was causing a name
collision with an enum member 'Type::SCOPE', which was therefore called
'SSCOPE'. As this is now unnecessary, it has been renamed to 'SCOPE'.